+++ Example/PrebidVeonDemo/src/main/java/org/prebid/veondemo/activities/ads/multi/MultiBannerActivity.kt
@@ -7,7 +7,7 @@ import android.widget.Toast
 import com.yandex.mobile.ads.common.ImpressionData
 import org.prebid.mobile.AdSize
 import org.prebid.mobile.api.data.SdkType
-import org.prebid.mobile.api.multiadloader.MultiBannerLoader
+import org.prebid.mobile.api.multiadloader.MultiBannerLoaderLegacyGam
 import org.prebid.mobile.api.multiadloader.listeners.MultiBannerViewListener
 import org.prebid.mobile.api.rendering.BannerView
 import org.prebid.veondemo.activities.BaseAdActivity
@@ -29,7 +29,7 @@ class MultiBannerActivity : BaseAdActivity() {
     }
 
     private fun createAd() {
-        val adLoader = MultiBannerLoader(
+        val adLoader = MultiBannerLoaderLegacyGam(
             context = this,
             adSize = AdSize(WIDTH, HEIGHT),
             configId = CONFIG_ID,