+++ Example/PrebidVeonDemo/src/main/java/org/prebid/veondemo/activities/ads/multi/MultiInterstitialAdActivity.kt
@@ -5,7 +5,7 @@ import android.util.Log
 import android.widget.Toast
 import com.yandex.mobile.ads.common.ImpressionData
 import org.prebid.mobile.api.data.SdkType
-import org.prebid.mobile.api.multiadloader.MultiInterstitialAdLoader
+import org.prebid.mobile.api.multiadloader.MultiInterstitialAdLoaderLegacyGam
 import org.prebid.mobile.api.multiadloader.listeners.MultiInterstitialAdListener
 import org.prebid.veondemo.activities.BaseAdActivity
 
@@ -24,7 +24,7 @@ class MultiInterstitialAdActivity : BaseAdActivity() {
     }
 
     private fun onCreateInterstitial() {
-        val interstitialLoader = MultiInterstitialAdLoader(
+        val interstitialLoader = MultiInterstitialAdLoaderLegacyGam(
             context = this,
             configId = CONFIG_ID,
             gamAdUnitId = AD_UNIT_ID,