+++ Example/PrebidVeonNextGenDemo/src/main/java/org/prebid/veonnextgendemo/NextGenApp.kt
@@ -1,4 +1,4 @@
-package org.prebid.nextgendemo
+package org.prebid.veonnextgendemo
 
 import android.app.Application
 import android.util.Log