+++ PrebidMobile/PrebidMobile-core/src/main/java/org/prebid/mobile/api/multiadloader/MultiInterstitialAdLoaderLegacyGam.kt
@@ -28,7 +28,7 @@ import org.prebid.mobile.logging.SdkAdStatus
 import org.prebid.mobile.logging.SdkLogUtil
 import java.util.EnumSet
 
-class MultiInterstitialAdLoader(
+class MultiInterstitialAdLoaderLegacyGam(
     private val context: Context,
     private val configId: String?,
     private val gamAdUnitId: String?,
@@ -36,7 +36,7 @@ class MultiInterstitialAdLoader(
 ) {
 
     companion object {
-        private val TAG = MultiInterstitialAdLoader::class.java.simpleName
+        private val TAG = MultiInterstitialAdLoaderLegacyGam::class.java.simpleName
     }
 
     enum class SdkState {